google-adk/self_sustained_mcp_agent/requirements.txt
litellm
google-adk
dotenv
google.genai